Coefficient of uniformity of germination
Source: R/CUGerm.R

Compute the coefficient of uniformity of germination (CUG).

CUGerm(germ.counts, intervals, partial = TRUE)

Arguments

germ.counts
Germination counts at each time interval. Can be partial or cumulative as specified in the argument partial.

intervals
The time intervals.

partial
logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

Value

The value of the coefficient of uniformity of germination as time-2.

Details

CUG is computed as follows (Heydecker, 1973, Bewley and Black, 1994).

CUG = ∑ki=1 Ni ⁄ ∑ki=1(T − Ti)2Ni

Where, T is the the mean germination time, Ti is the time from the start of the experiment to the ith observation (day for the example); Ni is the number of seeds germinated in the ith time (not the accumulated number, but the number correspondent to the ith observation), and k is the last time of germination.

References

Heydecker W (1972). Seed Ecology. Proceedings of the Nineteenth Easter School in Agricultural Science, University of Nottingham, 1972. Pennsylvania State University Press, University Park, USA.

Bewley JD and Black M (1994). Seeds: Physiology of Development and Germination.. Plenum Publishing Corporation, New York, USA. ISBN 0-306-44748-7, https://www.cabdirect.org/cabdirect/abstract/19950315483.

Examples

x <- c(0, 0, 0, 0, 4, 17, 10, 7, 1, 0, 1, 0, 0, 0)
y <- c(0, 0, 0, 0, 4, 21, 31, 38, 39, 39, 40, 40, 40, 40)
int <- 1:length(x)

# From partial germination counts
#----------------------------------------------------------------------------
CUGerm(germ.counts = x, intervals = int)

#> [1] 0.7092199

# From cumulative germination counts
#----------------------------------------------------------------------------
CUGerm(germ.counts = y, intervals = int, partial = FALSE)

#> [1] 0.05267935