Fit four-parameter hill function for multiple samples

Fit four-parameter hill function from a data frame of germination counts recorded at specific time intervals for multiple samples in batch.

FourPHFfit.bulk(data, total.seeds.col, counts.intervals.cols, intervals,
  partial = TRUE, fix.y0 = TRUE, fix.a = TRUE, tmax, xp = c(10,
  60), umin = 10, umax = 90, tries = 3)

Arguments

data
A data frame with the germination count data. It should possess columns with

Partial or cumulative germination counts per time interval (to be indicated by the argument counts.intervals.cols and

Total number of seeds tested (to be indicated by the argument total.seeds.col).

total.seeds.col
The name of the column in data with the total number of seeds tested.

counts.intervals.cols
The names of columns in data with the germination counts (partial or cumulative, as indicated by the argument partial) per time interval (indicated by the argument intervals).

intervals
The time intervals.

partial
logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

fix.y0
Force the intercept of the y axis through 0.

fix.a
Fix a as the actual maximum germination percentage at the end of the experiment.

tmax
The time up to which AUC is to be computed.

xp
Germination percentage value(s) for which the corresponding time is to be computed as a numeric vector. Default is c(10, 60).

umin
The minimum germination percentage value for computing uniformity. Default is 10. Seed Details.

umax
The maximum germination percentage value for computing uniformity. Default is 90. Seed Details.

tries
The number of tries to be attempted to fit the curve. Default is 3.

Value

A data frame with the original data along with the various parameters of the respective fitted four-parameter hill function.

References

El-Kassaby YA, Moss I, Kolotelo D, Stoehr M (2008). “Seed germination: Mathematical representation and parameters extraction.” Forest Science, 54(2), 220--227.

See also

Examples

data(gcdata)

counts.per.intervals <- c("Day01", "Day02", "Day03", "Day04", "Day05",
                          "Day06", "Day07", "Day08", "Day09", "Day10",
                          "Day11", "Day12", "Day13", "Day14")

FourPHFfit.bulk(gcdata, total.seeds.col = "Total Seeds",
                    counts.intervals.cols = counts.per.intervals,
                    intervals = 1:14, partial = TRUE,
                    fix.y0 = TRUE, fix.a = TRUE, xp = c(10, 60),
                    tmax = 20, tries = 3, umax = 90, umin = 10)

#> Error in FourPHFfit.bulk(gcdata, total.seeds.col = "Total Seeds", counts.intervals.cols = counts.per.intervals,     intervals = 1:14, partial = TRUE, fix.y0 = TRUE, fix.a = TRUE,     xp = c(10, 60), tmax = 20, tries = 3, umax = 90, umin = 10): could not find function "FourPHFfit.bulk"