Mean germination percentage and number of seeds per time interval
Source: R/MeanGermPercent.R

Compute the following metrics:

Mean/average germination percentage per unit time (GP)

Number of seeds germinated per unit time (N)

.

MeanGermPercent(germinated.seeds, germ.counts, total.seeds, intervals,
  partial = TRUE)

MeanGermNumber(germ.counts, intervals, partial = TRUE)

Arguments

germinated.seeds
Number of germinated seeds

germ.counts
Germination counts at each time interval. Can be partial or cumulative as specified in the argument partial.

total.seeds
Total number of seeds.

intervals
The time intervals.

partial
logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

Value

The value of mean germination percentage or mean number of seeds per time interval.

Details

Mean germination percentage per unit time (GP) is computed as follows (Czabator, 1962).

GP = GP ⁄ Tn

Where, GP is the final germination percentage and Tn is the total number of intervals(e.g. days) required for final germination.

Mean number of seeds germinated per unit time (N) is computed as follows (Khamassi et al., 2013).

N = Ng ⁄ Tn

Where, Ng is the number of germinated seeds and Tn is the total number of intervals (e.g. days) required for final germination.

References

Czabator FJ (1962). “Germination value: An index combining speed and completeness of pine seed germination.” Forest Science, 8(4), pp. 386--396. doi: 10.1093/forestscience/8.4.386 .

Djavanshir K and Pourbeik H (1976). “Germination value-A new formula.” Silvae genetica, 25(2), pp. 79--83.

Examples

x <- c(0, 0, 0, 0, 4, 17, 10, 7, 1, 0, 1, 0, 0, 0)
y <- c(0, 0, 0, 0, 4, 21, 31, 38, 39, 39, 40, 40, 40, 40)
int <- 1:length(x)

# From partial germination counts
#----------------------------------------------------------------------------
MeanGermPercent(germ.counts = x, total.seeds = 50, intervals = int)

#> [1] 5.714286

MeanGermNumber(germ.counts = x, intervals = int)

#> [1] 2.857143

# From cumulative germination counts
#----------------------------------------------------------------------------
MeanGermPercent(germ.counts = y, total.seeds = 50, intervals = int, partial = FALSE)

#> [1] 5.714286

MeanGermNumber(germ.counts = y, intervals = int, partial = FALSE)

#> [1] 2.857143

# From number of germinated seeds
#----------------------------------------------------------------------------
MeanGermPercent(germinated.seeds = 40, total.seeds = 50, intervals = int)

#> [1] 5.714286