Median germination time

Compute the median germination time (\(t_{50}\)). Median germination time is the time to reach 50% of final/maximum germination.

t50(germ.counts, intervals, partial = TRUE, method = c("coolbear", "farooq"))

Arguments

germ.counts

Germination counts at each time interval. Can be partial or cumulative as specified in the argument partial.

intervals

The time intervals.

partial

logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

method

The method for computing median germination time. Either "coolbear" or "farooq".

Value

The median germination time (\(t_{50}\)) value in the same unit of time as specified in the argument intervals.

Details

With argument method specified as "coolbear", median germination time is computed according to the formula by Coolbear et al. (1984) as follows.

\[t_{50}=T_{i}+ \frac{(\frac{N+1}{2}-N_{i})(T_{j}-T_{i})}{N_{j}-N_{i}}\]

Where, \(t_{50}\) is the median germination time, \(N\) is the final number of germinated seeds, and \(N_{i}\) and \(N_{j}\) are the total number of seeds germinated in adjacent counts at time \(T_{i}\) and \(T_{j}\) respectively, when \(N_{i} < \frac{N + 1}{2} < N_{j}\).

Similarly with argument method specified as "farooq", median germination time is computed according to the formula by by Farooq et al. (2005) as follows.

\[t_{50}=T_{i}+ \frac{(\frac{N}{2}-N_{i})(T_{j}-T_{i})}{N_{j}-N_{i}}\]

Where, \(t_{50}\) is the median germination time, \(N\) is the final number of germinated seeds, and \(N_{i}\) and \(N_{j}\) are the total number of seeds germinated in adjacent counts at time \(T_{i}\) and \(T_{j}\) respectively, when \(N_{i} < \frac{N}{2} < N_{j}\).

References

Coolbear P, Francis A, Grierson D (1984). “The effect of low temperature pre-sowing treatment on the germination performance and membrane integrity of artificially aged tomato seeds.” Journal of Experimental Botany, 35(11), 1609--1617.

Farooq M, Basra SMA, Ahmad N, Hafeez K (2005). “Thermal hardening: A new seed vigor enhancement tool in rice.” Journal of Integrative Plant Biology, 47(2), 187--193.

See also

Examples

x <- c(0, 0, 0, 0, 4, 17, 10, 7, 1, 0, 1, 0, 0, 0)
y <- c(0, 0, 0, 0, 4, 21, 31, 38, 39, 39, 40, 40, 40, 40)
int <- 1:length(x)

# From partial germination counts
#----------------------------------------------------------------------------
t50(germ.counts = x, intervals = int, method = "coolbear")
#> [1] 5.970588
t50(germ.counts = x, intervals = int, method = "farooq")
#> [1] 5.941176

# From cumulative germination counts
#----------------------------------------------------------------------------
t50(germ.counts = y, intervals = int, partial = FALSE, method = "coolbear")
#> [1] 5.970588
t50(germ.counts = y, intervals = int, partial = FALSE, method = "farooq")
#> [1] 5.941176