Fit four-parameter hill function for multiple samples

Fit four-parameter hill function from a data frame of germination counts recorded at specific time intervals for multiple samples in batch.

FourPHFfit.bulk(
  data,
  total.seeds.col,
  counts.intervals.cols,
  intervals,
  partial = TRUE,
  fix.y0 = TRUE,
  fix.a = TRUE,
  tmax,
  xp = c(10, 60),
  umin = 10,
  umax = 90,
  tries = 3
)

Arguments

data

A data frame with the germination count data. It should possess columns with

Partial or cumulative germination counts per time interval (to be indicated by the argument counts.intervals.cols and

Total number of seeds tested (to be indicated by the argument total.seeds.col).

total.seeds.col

The name of the column in data with the total number of seeds tested.

counts.intervals.cols

The names of columns in data with the germination counts (partial or cumulative, as indicated by the argument partial) per time interval (indicated by the argument intervals).

intervals

The time intervals.

partial

logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

fix.y0

Force the intercept of the y axis through 0.

fix.a

Fix a as the actual maximum germination percentage at the end of the experiment.

tmax

The time up to which AUC is to be computed.

xp

Germination percentage value(s) for which the corresponding time is to be computed as a numeric vector. Default is c(10, 60).

umin

The minimum germination percentage value for computing uniformity. Default is 10. Seed Details.

umax

The maximum germination percentage value for computing uniformity. Default is 90. Seed Details.

tries

The number of tries to be attempted to fit the curve. Default is 3.

Value

A data frame with the original data along with the various parameters of the respective fitted four-parameter hill function.

References

El-Kassaby YA, Moss I, Kolotelo D, Stoehr M (2008). “Seed germination: Mathematical representation and parameters extraction.” Forest Science, 54(2), 220--227.

See also

Examples

# \donttest{
data(gcdata)

counts.per.intervals <- c("Day01", "Day02", "Day03", "Day04", "Day05",
                          "Day06", "Day07", "Day08", "Day09", "Day10",
                          "Day11", "Day12", "Day13", "Day14")

FourPHFfit.bulk(gcdata, total.seeds.col = "Total Seeds",
                    counts.intervals.cols = counts.per.intervals,
                    intervals = 1:14, partial = TRUE,
                    fix.y0 = TRUE, fix.a = TRUE, xp = c(10, 60),
                    tmax = 20, tries = 3, umax = 90, umin = 10)
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> Warning: Convergence Error
#> #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#>    Genotype Rep Day01 Day02 Day03 Day04 Day05 Day06 Day07 Day08 Day09 Day10
#> 1        G1   1     0     0     0     0     4    17    10     7     1     0
#> 2        G2   1     0     0     0     1     3    15    13     6     2     1
#> 3        G3   1     0     0     0     2     3    18     9     8     2     1
#> 4        G4   1     0     0     0     0     4    19    12     6     2     1
#> 5        G5   1     0     0     0     0     5    20    12     8     1     0
#> 6        G1   2     0     0     0     0     3    21    11     7     1     1
#> 7        G2   2     0     0     0     0     4    18    11     7     1     0
#> 8        G3   2     0     0     0     1     3    14    12     6     2     1
#> 9        G4   2     0     0     0     1     3    19    10     8     1     1
#> 10       G5   2     0     0     0     0     4    18    13     6     2     1
#> 11       G1   3     0     0     0     0     5    21    11     8     1     0
#> 12       G2   3     0     0     0     0     3    20    10     7     1     1
#> 13       G3   3     0     0     0     0     4    19    12     8     1     1
#> 14       G4   3     0     0     0     0     3    21    11     6     1     0
#> 15       G5   3     0     0     0     0     4    17    10     8     1     1
#>    Day11 Day12 Day13 Day14 Total Seeds    a    b    c   y0  lag Dlag50
#> 1      1     0     0     0          50 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 2      0     1     0     0          51 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 3      1     1     0     0          48 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 4      1     1     0     0          51 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 5      0     1     1     0          50 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 6      1     1     0     0          49 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 7      1     0     0     0          48 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 8      0     1     0     0          47 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 9      1     1     0     0          52 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 10     0     1     0     0          50 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 11     0     1     1     0          51 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 12     1     1     0     0          51 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 13     0     1     1     0          49 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 14     1     1     0     0          48 <NA> <NA> <NA> <NA> <NA>   <NA>
#> 15     1     0     0     0          48 <NA> <NA> <NA> <NA> <NA>   <NA>
#>    t50.total txp.total_10 txp.total_60 t50.Germinated txp.Germinated_10
#> 1       <NA>         <NA>         <NA>           <NA>              <NA>
#> 2       <NA>         <NA>         <NA>           <NA>              <NA>
#> 3       <NA>         <NA>         <NA>           <NA>              <NA>
#> 4       <NA>         <NA>         <NA>           <NA>              <NA>
#> 5       <NA>         <NA>         <NA>           <NA>              <NA>
#> 6       <NA>         <NA>         <NA>           <NA>              <NA>
#> 7       <NA>         <NA>         <NA>           <NA>              <NA>
#> 8       <NA>         <NA>         <NA>           <NA>              <NA>
#> 9       <NA>         <NA>         <NA>           <NA>              <NA>
#> 10      <NA>         <NA>         <NA>           <NA>              <NA>
#> 11      <NA>         <NA>         <NA>           <NA>              <NA>
#> 12      <NA>         <NA>         <NA>           <NA>              <NA>
#> 13      <NA>         <NA>         <NA>           <NA>              <NA>
#> 14      <NA>         <NA>         <NA>           <NA>              <NA>
#> 15      <NA>         <NA>         <NA>           <NA>              <NA>
#>    txp.Germinated_60 Uniformity_90 Uniformity_10 Uniformity TMGR  AUC  MGT
#> 1               <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 2               <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 3               <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 4               <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 5               <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 6               <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 7               <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 8               <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 9               <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 10              <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 11              <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 12              <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 13              <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 14              <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#> 15              <NA>          <NA>          <NA>       <NA> <NA> <NA> <NA>
#>    Skewness
#> 1      <NA>
#> 2      <NA>
#> 3      <NA>
#> 4      <NA>
#> 5      <NA>
#> 6      <NA>
#> 7      <NA>
#> 8      <NA>
#> 9      <NA>
#> 10     <NA>
#> 11     <NA>
#> 12     <NA>
#> 13     <NA>
#> 14     <NA>
#> 15     <NA>
#>                                                                                              msg
#> 1  #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 2  #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 3  #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 4  #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 5  #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 6  #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 7  #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 8  #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 9  #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 10 #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 11 #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 12 #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 13 #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 14 #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#> 15 #1. failed to evaluate 'fn' at starting values: could not find function "FourPHF_fixa_fixy0" 
#>    Fit_sigma Fit_isConv Fit_finTol Fit_logLik Fit_AIC Fit_BIC Fit_deviance
#> 1       <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 2       <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 3       <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 4       <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 5       <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 6       <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 7       <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 8       <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 9       <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 10      <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 11      <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 12      <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 13      <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 14      <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#> 15      <NA>       <NA>       <NA>       <NA>    <NA>    <NA>         <NA>
#>    Fit_df.residual Fit_nobs
#> 1             <NA>     <NA>
#> 2             <NA>     <NA>
#> 3             <NA>     <NA>
#> 4             <NA>     <NA>
#> 5             <NA>     <NA>
#> 6             <NA>     <NA>
#> 7             <NA>     <NA>
#> 8             <NA>     <NA>
#> 9             <NA>     <NA>
#> 10            <NA>     <NA>
#> 11            <NA>     <NA>
#> 12            <NA>     <NA>
#> 13            <NA>     <NA>
#> 14            <NA>     <NA>
#> 15            <NA>     <NA>
# }