Fit four-parameter hill function for multiple samples

Fit four-parameter hill function from a data frame of germination counts recorded at specific time intervals for multiple samples in batch.

FourPHFfit.bulk(
  data,
  total.seeds.col,
  counts.intervals.cols,
  intervals,
  partial = TRUE,
  fix.y0 = TRUE,
  fix.a = TRUE,
  tmax,
  xp = c(10, 60),
  umin = 10,
  umax = 90,
  tries = 3
)

Arguments

data

A data frame with the germination count data. It should possess columns with

Partial or cumulative germination counts per time interval (to be indicated by the argument counts.intervals.cols and

Total number of seeds tested (to be indicated by the argument total.seeds.col).

total.seeds.col

The name of the column in data with the total number of seeds tested.

counts.intervals.cols

The names of columns in data with the germination counts (partial or cumulative, as indicated by the argument partial) per time interval (indicated by the argument intervals).

intervals

The time intervals.

partial

logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

fix.y0

Force the intercept of the y axis through 0.

fix.a

Fix a as the actual maximum germination percentage at the end of the experiment.

tmax

The time up to which AUC is to be computed.

xp

Germination percentage value(s) for which the corresponding time is to be computed as a numeric vector. Default is c(10, 60).

umin

The minimum germination percentage value for computing uniformity. Default is 10. Seed Details.

umax

The maximum germination percentage value for computing uniformity. Default is 90. Seed Details.

tries

The number of tries to be attempted to fit the curve. Default is 3.

Value

A data frame with the original data along with the various parameters of the respective fitted four-parameter hill function.

References

El-Kassaby YA, Moss I, Kolotelo D, Stoehr M (2008). “Seed germination: Mathematical representation and parameters extraction.” Forest Science, 54(2), 220--227.

See also

Examples

# \donttest{
data(gcdata)

counts.per.intervals <- c("Day01", "Day02", "Day03", "Day04", "Day05",
                          "Day06", "Day07", "Day08", "Day09", "Day10",
                          "Day11", "Day12", "Day13", "Day14")

FourPHFfit.bulk(gcdata, total.seeds.col = "Total Seeds",
                    counts.intervals.cols = counts.per.intervals,
                    intervals = 1:14, partial = TRUE,
                    fix.y0 = TRUE, fix.a = TRUE, xp = c(10, 60),
                    tmax = 20, tries = 3, umax = 90, umin = 10)
#>    Genotype Rep Day01 Day02 Day03 Day04 Day05 Day06 Day07 Day08 Day09 Day10
#> 1        G1   1     0     0     0     0     4    17    10     7     1     0
#> 2        G2   1     0     0     0     1     3    15    13     6     2     1
#> 3        G3   1     0     0     0     2     3    18     9     8     2     1
#> 4        G4   1     0     0     0     0     4    19    12     6     2     1
#> 5        G5   1     0     0     0     0     5    20    12     8     1     0
#> 6        G1   2     0     0     0     0     3    21    11     7     1     1
#> 7        G2   2     0     0     0     0     4    18    11     7     1     0
#> 8        G3   2     0     0     0     1     3    14    12     6     2     1
#> 9        G4   2     0     0     0     1     3    19    10     8     1     1
#> 10       G5   2     0     0     0     0     4    18    13     6     2     1
#> 11       G1   3     0     0     0     0     5    21    11     8     1     0
#> 12       G2   3     0     0     0     0     3    20    10     7     1     1
#> 13       G3   3     0     0     0     0     4    19    12     8     1     1
#> 14       G4   3     0     0     0     0     3    21    11     6     1     0
#> 15       G5   3     0     0     0     0     4    17    10     8     1     1
#>    Day11 Day12 Day13 Day14 Total Seeds                a                b
#> 1      1     0     0     0          50               80 9.88193689239939
#> 2      0     1     0     0          51 82.3529411764706 9.22766646359687
#> 3      1     1     0     0          48            93.75 7.79305096503615
#> 4      1     1     0     0          51 90.1960784313725 8.92565503394839
#> 5      0     1     1     0          50               96  9.4191816695981
#> 6      1     1     0     0          49 93.8775510204082 9.45014899129514
#> 7      1     0     0     0          48             87.5 10.1724586100529
#> 8      0     1     0     0          47 85.1063829787234 8.94069602989349
#> 9      1     1     0     0          52 86.5384615384615  8.6173913532163
#> 10     0     1     0     0          50               90 9.60884373087692
#> 11     0     1     1     0          51 94.1176470588235 9.40021183872586
#> 12     1     1     0     0          51 86.2745098039216 9.16252658054406
#> 13     0     1     1     0          49 95.9183673469388 8.99520959277319
#> 14     1     1     0     0          48 91.6666666666667 10.3918447990981
#> 15     1     0     0     0          48             87.5 9.13674439831543
#>                   c y0 lag           Dlag50        t50.total     txp.total_10
#> 1  6.03495355423622  0   0 6.03495355423622 6.35512149738159 4.95626430996844
#> 2  6.17519294875847  0   0 6.17519294875847 6.47349043979169 4.98323617961634
#> 3  6.13811027323845  0   0 6.13811027323845 6.24419102980473 4.67302155333174
#> 4  6.12517308176588  0   0 6.12517308176588 6.27679437746254 4.85087548237175
#> 5  6.04964210720327  0   0 6.04964210720327 6.10343321091848 4.81412549010201
#> 6  6.09741485213496  0   0 6.09741485213496 6.18227860747235 4.86863251431733
#> 7  6.02985089631599  0   0 6.02985089631599 6.20281219696422 4.93042184740182
#> 8  6.18977354961439  0   0 6.18977354961439 6.43951015764455 4.94005695310539
#> 9  6.12512151399929  0   0 6.12512151399929 6.35217197764166 4.83665841861718
#> 10 6.10950363575767  0   0 6.10950363575767 6.25304320794668 4.92062915221628
#> 11 6.01875974061195  0   0 6.01875974061195 6.09943499335382 4.79862683383817
#> 12  6.1084516820797  0   0  6.1084516820797 6.32618435705024 4.89359557090626
#> 13 6.14901168803061  0   0 6.14901168803061 6.20750091288263 4.84130798267247
#> 14 6.01591019543247  0   0 6.01591019543247 6.12238872898609 4.91514013767951
#> 15 6.12157936163499  0   0 6.12157936163499 6.31739163301497 4.89250226946576
#>        txp.total_60   t50.Germinated txp.Germinated_10 txp.Germinated_60
#> 1  6.74459834631771 6.03495355423622   4.8318073794034  6.28772357367188
#> 2  6.87260337968692 6.17519294875847  4.86675518549505  6.45258151256586
#> 3  6.60843809251509 6.13811027323845  4.63006208020014  6.46592435698387
#> 4  6.61496814302537 6.12517308176588  4.78859693817119  6.40983765941072
#> 5  6.38678874941426 6.04964210720327  4.79094574322756  6.31574586639992
#> 6  6.47759860932629 6.09741485213496  4.83247140619904  6.36472210249767
#> 7     6.51049505523 6.02985089631599  4.85847638047658   6.2750496018235
#> 8  6.82329908278267 6.18977354961439  4.84110536088622  6.47694540370958
#> 9  6.73327569782723 6.12512151399929  4.74657350251934  6.42020821882777
#> 10  6.5665061956458 6.10950363575767  4.86068135560336  6.37282341572477
#> 11  6.3912906236839 6.01875974061195  4.76424552194859   6.2840509537431
#> 12 6.68452626570581  6.1084516820797  4.80601279742022  6.38483647023757
#> 13 6.50995387022312  6.1490116880306  4.81639290881553  6.43252427366119
#> 14 6.39749097966013 6.01591019543247  4.86939775646343   6.2552761045868
#> 15 6.66724718740801 6.12157936163499  4.81308335438754  6.39935718177504
#>       Uniformity_90    Uniformity_10       Uniformity             TMGR
#> 1  7.53768963494679  4.8318073794034 2.70588225554338 5.91219440465559
#> 2  7.83540706301571 4.86675518549505 2.96865187752066 6.03128155417137
#> 3  8.13734180531911 4.63006208020014 3.50727972511896 5.93817948829363
#> 4  7.83480960415051 4.78859693817119 3.04621266597932 5.97268622562109
#> 5  7.63902819750811 4.79094574322756 2.84808245428055 5.91428884333636
#> 6  7.69346877693759 4.83247140619904 2.86099737073854 5.96187868562603
#> 7  7.48364280989593 4.85847638047658 2.62516642941935 5.91405695229978
#> 8  7.91416293168472 4.84110536088622 3.07305757079851 6.03619216805867
#> 9  7.90404141879274 4.74657350251934  3.1574679162734  5.9616310497804
#> 10 7.67917745365204 4.86068135560336 2.81849609804867 5.97811533002658
#> 11 7.60361082322955 4.76424552194859 2.83936530128096 5.88355748786772
#> 12 7.76385405638773 4.80601279742022 2.95784125896751  5.9640804983933
#> 13 7.85034474042432 4.81639290881553 3.03395183160878 5.99827012388754
#> 14 7.43237198716613 4.86939775646343  2.5629742307027 5.90518049089766
#> 15 7.78580612916975 4.81308335438754 2.97272277478221 5.97608676470078
#>                 AUC              MGT         Skewness          msg
#> 1  1108.97550938753  6.6322519662712 1.09897315806444 #1. success 
#> 2  1128.55880088539 6.78440735640875 1.09865512101493 #1. success 
#> 3  1283.69307344325  6.7727423279724 1.10339209080373 #1. success 
#> 4  1239.88674124826 6.73966592721389 1.10032252758331 #1. success 
#> 5  1328.32820017628 6.65498075748102 1.10006189449736 #1. success 
#> 6  1294.46271443954 6.70247312592894  1.0992319349211 #1. success 
#> 7  1213.90764565674 6.62241708548249 1.09827211308468 #1. success 
#> 8  1164.34586106316 6.80400021213917 1.09923249333783 #1. success 
#> 9  1188.79304149759  6.7452410863068 1.10124200326315 #1. success 
#> 10 1240.22733172773 6.71189998816383 1.09859988442931 #1. success 
#> 11 1305.20007906005 6.62424817630914 1.10060020033889 #1. success 
#> 12  1188.0211599463 6.71863893649018 1.09989229450739 #1. success 
#> 13 1316.40687295539 6.76227360647556 1.09973341238539 #1. success 
#> 14 1273.38526597424 6.60496678828807  1.0979164538232 #1. success 
#> 15 1203.66421628837 6.73226579042194 1.09975961965212 #1. success 
#>           Fit_sigma Fit_isConv           Fit_finTol        Fit_logLik
#> 1  1.61522002910957       TRUE  2.8528290840768e-12  -25.498681342686
#> 2  1.11537185901124       TRUE 5.11413134063332e-12 -20.3147146781893
#> 3  2.43270386985341       TRUE 8.43982661535847e-11 -31.2321314996742
#> 4  2.39658164351394       TRUE 3.38218342221808e-12 -31.0226924019787
#> 5  2.39966172990826       TRUE 6.74447164783487e-11 -31.0406736477542
#> 6  3.03496223650969       TRUE 3.95630195271224e-11  -34.328870450832
#> 7  1.66301938705135       TRUE 3.90798504668055e-12 -25.9069727183683
#> 8  1.12070433595621       TRUE 4.32720526077901e-12 -20.3814877326307
#> 9  2.42996010854989       TRUE 1.77209358298569e-11 -31.2163324798379
#> 10 1.68665620116432       TRUE 8.19966317067156e-12 -26.1045565628368
#> 11 2.62811272107047       TRUE 1.32729383039987e-11 -32.3138085946749
#> 12 2.87814601795845       TRUE 3.51434437106946e-11 -33.5861335093548
#> 13 2.60458797517776       TRUE 1.00897068477934e-11 -32.1879276469568
#> 14 2.76475621724483       TRUE 9.80548975348938e-13  -33.023419198233
#> 15 1.95400807212262       TRUE 8.73967564984923e-13 -28.1644422917083
#>             Fit_AIC          Fit_BIC     Fit_deviance Fit_df.residual Fit_nobs
#> 1  56.9973626853719 58.9145346742177 31.3072289092405              12       14
#> 2  46.6294293563787 48.5466013452244 14.9286526064904              12       14
#> 3  68.4642629993484 70.3814349881942 71.0165774207973              12       14
#> 4  68.0453848039574 69.9625567928032  68.923242888336              12       14
#> 5  68.0813472955084 69.9985192843541 69.1005170158358              12       14
#> 6  74.6577409016639 76.5749128905097 110.531949324479              12       14
#> 7  57.8139454367367 59.7311174255824 33.1876017805038              12       14
#> 8  46.7629754652615 48.6801474541073 15.0717385035725              12       14
#> 9  68.4326649596759 70.3498369485217 70.8564735497253              12       14
#> 10 58.2091131256735 60.1262851145193 34.1377096911126              12       14
#> 11 70.6276171893498 72.5447891781956 82.8837176958294              12       14
#> 12 73.1722670187096 75.0894390075554 99.4046940082808              12       14
#> 13 70.3758552939136 72.2930272827594 81.4065422452872              12       14
#> 14 72.0468383964661 73.9640103853119 91.7265232895271              12       14
#> 15 62.3288845834165 64.2460565722623 45.8177705510444              12       14
# }