Fit four-parameter hill function for multiple samples

Fit four-parameter hill function from a data frame of germination counts recorded at specific time intervals for multiple samples in batch.

FourPHFfit.bulk(
  data,
  total.seeds.col,
  counts.intervals.cols,
  intervals,
  partial = TRUE,
  fix.y0 = TRUE,
  fix.a = TRUE,
  tmax,
  xp = c(10, 60),
  umin = 10,
  umax = 90,
  tries = 3
)

Arguments

data

A data frame with the germination count data. It should possess columns with

Partial or cumulative germination counts per time interval (to be indicated by the argument counts.intervals.cols and

Total number of seeds tested (to be indicated by the argument total.seeds.col).

total.seeds.col

The name of the column in data with the total number of seeds tested.

counts.intervals.cols

The names of columns in data with the germination counts (partial or cumulative, as indicated by the argument partial) per time interval (indicated by the argument intervals).

intervals

The time intervals.

partial

logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

fix.y0

Force the intercept of the y axis through 0.

fix.a

Fix a as the actual maximum germination percentage at the end of the experiment.

tmax

The time up to which AUC is to be computed.

xp

Germination percentage value(s) for which the corresponding time is to be computed as a numeric vector. Default is c(10, 60).

umin

The minimum germination percentage value for computing uniformity. Default is 10. Seed Details.

umax

The maximum germination percentage value for computing uniformity. Default is 90. Seed Details.

tries

The number of tries to be attempted to fit the curve. Default is 3.

Value

A data frame with the original data along with the various parameters of the respective fitted four-parameter hill function.

References

El-Kassaby YA, Moss I, Kolotelo D, Stoehr M (2008). “Seed germination: Mathematical representation and parameters extraction.” Forest Science, 54(2), 220--227.

See also

Examples

# \donttest{
data(gcdata)

counts.per.intervals <- c("Day01", "Day02", "Day03", "Day04", "Day05",
                          "Day06", "Day07", "Day08", "Day09", "Day10",
                          "Day11", "Day12", "Day13", "Day14")

FourPHFfit.bulk(gcdata, total.seeds.col = "Total Seeds",
                    counts.intervals.cols = counts.per.intervals,
                    intervals = 1:14, partial = TRUE,
                    fix.y0 = TRUE, fix.a = TRUE, xp = c(10, 60),
                    tmax = 20, tries = 3, umax = 90, umin = 10)
#>    Genotype Rep Day01 Day02 Day03 Day04 Day05 Day06 Day07 Day08 Day09 Day10
#> 1        G1   1     0     0     0     0     4    17    10     7     1     0
#> 2        G2   1     0     0     0     1     3    15    13     6     2     1
#> 3        G3   1     0     0     0     2     3    18     9     8     2     1
#> 4        G4   1     0     0     0     0     4    19    12     6     2     1
#> 5        G5   1     0     0     0     0     5    20    12     8     1     0
#> 6        G1   2     0     0     0     0     3    21    11     7     1     1
#> 7        G2   2     0     0     0     0     4    18    11     7     1     0
#> 8        G3   2     0     0     0     1     3    14    12     6     2     1
#> 9        G4   2     0     0     0     1     3    19    10     8     1     1
#> 10       G5   2     0     0     0     0     4    18    13     6     2     1
#> 11       G1   3     0     0     0     0     5    21    11     8     1     0
#> 12       G2   3     0     0     0     0     3    20    10     7     1     1
#> 13       G3   3     0     0     0     0     4    19    12     8     1     1
#> 14       G4   3     0     0     0     0     3    21    11     6     1     0
#> 15       G5   3     0     0     0     0     4    17    10     8     1     1
#>    Day11 Day12 Day13 Day14 Total Seeds        a         b        c y0 lag
#> 1      1     0     0     0          50 80.00000  9.881947 6.034954  0   0
#> 2      0     1     0     0          51 82.35294  9.227667 6.175193  0   0
#> 3      1     1     0     0          48 93.75000  7.793055 6.138110  0   0
#> 4      1     1     0     0          51 90.19608  8.925668 6.125172  0   0
#> 5      0     1     1     0          50 96.00000  9.419194 6.049641  0   0
#> 6      1     1     0     0          49 93.87755  9.450187 6.097412  0   0
#> 7      1     0     0     0          48 87.50000 10.172466 6.029851  0   0
#> 8      0     1     0     0          47 85.10638  8.940702 6.189774  0   0
#> 9      1     1     0     0          52 86.53846  8.617395 6.125121  0   0
#> 10     0     1     0     0          50 90.00000  9.608849 6.109503  0   0
#> 11     0     1     1     0          51 94.11765  9.400248 6.018759  0   0
#> 12     1     1     0     0          51 86.27451  9.162558 6.108449  0   0
#> 13     0     1     1     0          49 95.91837  8.995233 6.149011  0   0
#> 14     1     1     0     0          48 91.66667 10.391898 6.015907  0   0
#> 15     1     0     0     0          48 87.50000  9.136762 6.121580  0   0
#>      Dlag50 t50.total t50.Germinated     TMGR      AUC      MGT Skewness
#> 1  6.034954  6.355122       6.034954 5.912195 1108.975 6.632252 1.098973
#> 2  6.175193  6.473490       6.175193 6.031282 1128.559 6.784407 1.098655
#> 3  6.138110  6.244190       6.138110 5.938179 1283.693 6.772742 1.103392
#> 4  6.125172  6.276793       6.125172 5.972686 1239.887 6.739665 1.100323
#> 5  6.049641  6.103433       6.049641 5.914289 1328.328 6.654980 1.100062
#> 6  6.097412  6.182276       6.097412 5.961877 1294.463 6.702470 1.099232
#> 7  6.029851  6.202812       6.029851 5.914057 1213.908 6.622417 1.098272
#> 8  6.189774  6.439510       6.189774 6.036193 1164.346 6.804000 1.099232
#> 9  6.125121  6.352172       6.125121 5.961631 1188.793 6.745241 1.101242
#> 10 6.109503  6.253042       6.109503 5.978115 1240.227 6.711899 1.098600
#> 11 6.018759  6.099434       6.018759 5.883558 1305.200 6.624247 1.100600
#> 12 6.108449  6.326181       6.108449 5.964079 1188.021 6.718636 1.099892
#> 13 6.149011  6.207500       6.149011 5.998270 1316.407 6.762272 1.099733
#> 14 6.015907  6.122385       6.015907 5.905179 1273.386 6.604963 1.097916
#> 15 6.121580  6.317392       6.121580 5.976088 1203.664 6.732267 1.099760
#>                                                             msg isConv
#> 1  #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 2  #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 3  #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 4  #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 5  #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 6  #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 7  #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 8  #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 9  #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 10 #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 11 #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 12 #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 13 #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 14 #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#> 15 #1. Relative error in the sum of squares is at most `ftol'.    TRUE
#>    txp.total_10 txp.total_60 Uniformity_90 Uniformity_10 Uniformity
#> 1      4.956266     6.744598      7.537688      4.831809   2.705880
#> 2      4.983236     6.872603      7.835407      4.866755   2.968652
#> 3      4.673022     6.608437      8.137340      4.630062   3.507277
#> 4      4.850876     6.614967      7.834806      4.788598   3.046208
#> 5      4.814126     6.386788      7.639025      4.790947   2.848078
#> 6      4.868635     6.477594      7.693458      4.832474   2.860984
#> 7      4.930423     6.510495      7.483642      4.858477   2.625165
#> 8      4.940058     6.823299      7.914162      4.841106   3.073056
#> 9      4.836659     6.733275      7.904040      4.746574   3.157466
#> 10     4.920629     6.566505      7.679176      4.860681   2.818494
#> 11     4.798630     6.391288      7.603603      4.764249   2.839354
#> 12     4.893597     6.684521      7.763844      4.806015   2.957830
#> 13     4.841310     6.509952      7.850339      4.816395   3.033943
#> 14     4.915143     6.397486      7.432360      4.869401   2.562960
#> 15     4.892505     6.667247      7.785804      4.813086   2.972718
# }