Speed of germination
Source: R/GermSpeed.R

Compute the following metrics:

Speed of germination or Germination rate index or Index of velocity of germination or Germination index (Kendrick and Frankland, 1969; AOSA, 1983)

Speed of accumulated germination (Bradbeer, 1988; Wardle et al., 1991; Haugland and Brandsaeter, 1996; Santana and Ranal, 2004)

Corrected speed of germination or Corrected germination rate index

GermSpeed(germ.counts, intervals, partial = TRUE)

GermSpeedAccumulated(germ.counts, intervals, partial = TRUE)

GermSpeedCorrected(germ.counts, intervals, partial = TRUE, total.seeds)

Arguments

germ.counts
Germination counts at each time interval. Can be partial or cumulative as specified in the argument partial.

intervals
The time intervals.

partial
logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

total.seeds
Total number of seeds.

Value

For GermSpeed, the value of germination speed as % time-1.

For GermSpeedAccumulated, the value of accumulated germination speed as % time-1.

For GermSpeedCorrected, the corrected speed of germination.

Details

GermSpeed computes the speed of germination according to the following formula (Throneberry and Smith, 1955; Maguire, 1962; Kendrick and Frankland, 1969; AOSA, 1983; Khandakar and Bradbeer, 1983; Bradbeer, 1988; Wardle et al., 1991).

S = N1⁄T1+N2⁄T2+N3⁄T3+ …+Nn⁄Tn

Where, N1, N2, N3, …, Nn are the number of germinated seeds observed at time (days or hours) T1, T2, T3, …, Tn after sowing. (Not accumulated/cumulative number, but the number of seeds that germinated at the specific time).

Speed of germination expresses the rate of germination in terms of the total number of seeds that germinate in a time interval. Higher values indicate greater and faster germination. This is useful for comparisons only when samples or treatments possess similar germinabilities. This is overcome by using the corrected speed of germination.

GermSpeedCorrected computes the corrected speed of germination as follows (Evetts and Burnside, 1972).

Scorrected = S⁄FGP

Where, FGP is the final germination percentage or germinability.

GermSpeedAccumulated computes the speed of accumulated germination as follows (Bradbeer, 1988; Wardle et al., 1991; Haugland and Brandsaeter, 1996; Santana and Ranal, 2004).

S = N1⁄T1 + [N1+N2]⁄T2 + [N1+N2+N3]⁄T3 + …+ [N1+N2+ …+Nn]⁄Tn

Where, N1, N2, N3, …, Nn are the number of germinated seeds observed at time (days or hours) T1, T2, T3, …, Tn after sowing. (Not accumulated/cumulative number, but the number of seeds that germinated at the specific time).

References

Throneberry GO and Smith FG (1955). “Relation of respiratory and enzymatic activity to corn seed viability.” Plant Physiology, 30(4), pp. 337--343.

Maguire JD (1962). “Speed of germination - Aid in selection and evaluation for seedling emergence and vigor.” Crop Science, 2(2), pp. 176--177. doi: 10.2135/cropsci1962.0011183x000200020033x .

Kendrick RE and Frankland B (1969). “Photocontrol of germination in Amaranthus caudatus.” Planta, 85(4), pp. 326--339. doi: 10.1007/bf00381281 .

Evetts LL and Burnside OC (1972). “Germination and seedling development of common milkweed and other species.” Weed Science, 20(4), pp. 371--378. doi: 10.1017/S004317450003589X .

Khandakar A and Bradbeer J (1983). “Jute seed quality.” Bangladesh Agricultural Research Council, Dhaka.

Bradbeer JW (1988). Seed Dormancy and Germination. Blackie, Glasgow and London. www.springer.com/in/book/9780216916364.

Wardle D, Ahmed M and Nicholson KS (1991). “Allelopathic influence of nodding thistle (Carduus nutans L.) seeds on germination and radicle growth of pasture plants.” New Zealand Journal of Agricultural Research, 34(2), pp. 185--191. doi: 10.1080/00288233.1991.10423358 .

Haugland E and Brandsaeter LO (1996). “Experiments on bioassay sensitivity in the study of allelopathy.” Journal of Chemical Ecology, 22(10), pp. 1845--1859.

de Santana D and Ranal MA (2004). Analise Da Germinacao: Um Enfoque Estatistico. Universidade de Brasilia, Brasilia.

Examples

x <- c(0, 0, 0, 0, 4, 17, 10, 7, 1, 0, 1, 0, 0, 0)
y <- c(0, 0, 0, 0, 4, 21, 31, 38, 39, 39, 40, 40, 40, 40)
int <- 1:length(x)

# From partial germination counts
#----------------------------------------------------------------------------
GermSpeed(germ.counts = x, intervals = int)

#> [1] 6.138925

GermSpeedAccumulated(germ.counts = x, intervals = int)

#> [1] 34.61567

GermSpeedCorrected(germ.counts = x, intervals = int, total.seeds = 50)

#> [1] 0.07673656

# From cumulative germination counts
#----------------------------------------------------------------------------
GermSpeed(germ.counts = y, intervals = int, partial = FALSE)

#> [1] 6.138925

GermSpeedAccumulated(germ.counts = y, intervals = int, partial = FALSE)

#> [1] 34.61567

GermSpeedCorrected(germ.counts = y, intervals = int,
                   partial = FALSE, total.seeds = 50)

#> [1] 0.009245369