Synchrony and uncertainty of germination
Source: R/GermSynchrony.R

Compute the following metrics:

GermSynchrony

Synchrony of germination (Z index).

GermUncertainty

Synchronization index (E) or Uncertainty of the germination process (U) or Informational entropy (H).

GermSynchrony(germ.counts, intervals, partial = TRUE)

GermUncertainty(germ.counts, intervals, partial = TRUE)

Arguments

germ.counts
Germination counts at each time interval. Can be partial or cumulative as specified in the argument partial.

intervals
The time intervals.

partial
logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

Value

For GermUncertainty, the value of uncertainty of germination process.

For GermSynchrony, the value of synchrony of germination.

Details

GermSynchrony computes the value of synchrony of germination (Z index) as follows (Primack, 1980; Ranal and Santana, 2006).

Z = ∑ki=1CNi,2 ⁄C∑Ni,2

Where, CNi,2 is the partial combination of the two germinated seeds from among Ni, the number of seeds germinated on the ith time (estimated as CNi,2 = [Ni(Ni-1)] ⁄ 2) and C∑Ni,2 is the partial combination of the two germinated seeds from among the total number of seeds germinated at the final count, assuming that all seeds that germinated did so simultaneously.

GermUncertainty computes the value of synchronization index (E) or uncertainty of the germination process (U) or informational entropy (H) as follows (Shannon, 1948; Labouriau and Valadares, 1976; Labouriau 1983).

E = −∑ki=1 filog 2fi

Where, fi is the relative frequency of germination (estimated as fi = Ni ⁄ ∑ki=1Ni), Ni is the number of seeds germinated on the ith time and k is the last day of observation.

References

Shannon CE (1948). “A mathematical theory of communication.” Bell System Technical Journal, 27(3), pp. 379--423. doi: 10.1002/j.1538-7305.1948.tb01338.x .

Labouriau L and Viladares M (1976). “On the germination of seeds of Calotropis procera (Ait.) Ait. f.” Anais da Academia Brasileira de Ciencias, 48(263-284).

Labouriau L (1983). “Uma nova linha de pesquisa na fisiologia da germinacao das sementes.” Anais do XXXIV Congresso Nacional de Botanica. SBB, Porto Alegre, pp. 11--50.

Primack RB (1985). “Longevity of individual flowers.” Annual Review of Ecology and Systematics, 16(1), pp. 15--37. doi: 10.1146/annurev.es.16.110185.000311 .

Ranal MA and de Santana DG (2006). “How and why to measure the germination process?” Brazilian Journal of Botany, 29(1), pp. 1--11. doi: 10.1590/s0100-84042006000100002 .

See also

Examples

x <- c(0, 0, 0, 0, 4, 17, 10, 7, 1, 0, 1, 0, 0, 0)
y <- c(0, 0, 0, 0, 4, 21, 31, 38, 39, 39, 40, 40, 40, 40)
int <- 1:length(x)

# From partial germination counts
#----------------------------------------------------------------------------
GermSynchrony(germ.counts = x, intervals = int)

#> [1] 0.2666667

GermUncertainty(germ.counts = x, intervals = int)

#> [1] 2.062987

# From cumulative germination counts
#----------------------------------------------------------------------------
GermSynchrony(germ.counts = y, intervals = int, partial = FALSE)

#> [1] 0.2666667

GermUncertainty(germ.counts = y, intervals = int, partial = FALSE)

#> [1] 2.062987