Peak value and germination value

Compute the peak value (Czabator, 1962) and germination value (Czabator, 1962; Djavanshir and Pourbeik, 1976).

PeakValue(germ.counts, intervals, total.seeds, partial = TRUE)

GermValue(germ.counts, intervals, total.seeds, partial = TRUE,
  method = c("czabator", "dp"), k = 10)

Arguments

germ.counts
Germination counts at each time interval. Can be partial or cumulative as specified in the argument partial.

intervals
The time intervals.

total.seeds
Total number of seeds.

partial
logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

method
The method for computing germination value. Either "czabator" or "dp".

k
Constant (See Details). Default is 10.

Value

A list with the following components:

Germination Value

The germination value

Calculations

The data frame of calculations.

testend

The end of test value (Only for method dp).

Details

Peak value (PV) is the maximum quotient obtained by dividing successive cumulative germination values by the relevant incubation time (Czabator, 1962). It represents the mean daily germination of the most vigorous component of the seed lot, and is a mathematical expression of the break, or shoulder, of a typical sigmoid germination curve (Djavanshir and Pourbeik, 1976).

For daily germination counts, germination value (GV) is computed as follows (Czabator, 1962).

GV = PV × MDG

Where, PV is the peak value and MDG is the mean daily germination percentage.

GV can also be computed for other time intervals of successive germination counts, by replacing MDG with the mean germination percentage per unit time (GP).

A new estimation of germination value was given by Djavanshir and Pourbeik (1976) as follows:

GV = [∑DGS ⁄ N] × GP × k

Where, DGS is the daily germination speed computed by dividing cumulative germination percentage by the number of days since the beginning of the test, N is the frequency or number of DGS calculated during the test, \(GP\) is the germination percentage expressed over 100 and k is a constant. The value of k is decided on the basis of average daily speed of germination (∑DGS ⁄ N). If it is less than 10, then k value of 10 can be used and if it is more than 10, then value of 7 or 8 can be used for k.

References

Czabator FJ (1962). “Germination value: An index combining speed and completeness of pine seed germination.” Forest Science, 8(4), 386--396.

Khamassi K, Harbaoui K, Jaime ATdS, Jeddi FB (2013). “Optimal germination temperature assessed by indices and models in field bean (Vicia faba L. var. minor).” Agriculturae Conspectus Scientificus, 78(2), 131--136. https://hrcak.srce.hr/104663.

Examples

x <- c(0, 0, 34, 40, 21, 10, 4, 5, 3, 5, 8, 7, 7, 6, 6, 4, 0, 2, 0, 2)
y <- c(0, 0, 34, 74, 95, 105, 109, 114, 117, 122, 130, 137, 144, 150,
      156, 160, 160, 162, 162, 164)
int <- 1:length(x)
total.seeds = 200

# From partial germination counts
#----------------------------------------------------------------------------
PeakValue(germ.counts = x, intervals = int, total.seeds = 200)

#> [1] 9.5

GermValue(germ.counts = x, intervals = int, total.seeds = 200,
          method = "czabator")

#> $`Germination Value`
#> [1] 38.95
#> 
#> [[2]]
#>    germ.counts intervals Cumulative.germ.counts Cumulative.germ.percent
#> 3           34         3                     34                    17.0
#> 4           40         4                     74                    37.0
#> 5           21         5                     95                    47.5
#> 6           10         6                    105                    52.5
#> 7            4         7                    109                    54.5
#> 8            5         8                    114                    57.0
#> 9            3         9                    117                    58.5
#> 10           5        10                    122                    61.0
#> 11           8        11                    130                    65.0
#> 12           7        12                    137                    68.5
#> 13           7        13                    144                    72.0
#> 14           6        14                    150                    75.0
#> 15           6        15                    156                    78.0
#> 16           4        16                    160                    80.0
#> 17           0        17                    160                    80.0
#> 18           2        18                    162                    81.0
#> 19           0        19                    162                    81.0
#> 20           2        20                    164                    82.0
#>         DGS
#> 3  5.666667
#> 4  9.250000
#> 5  9.500000
#> 6  8.750000
#> 7  7.785714
#> 8  7.125000
#> 9  6.500000
#> 10 6.100000
#> 11 5.909091
#> 12 5.708333
#> 13 5.538462
#> 14 5.357143
#> 15 5.200000
#> 16 5.000000
#> 17 4.705882
#> 18 4.500000
#> 19 4.263158
#> 20 4.100000
#> 

GermValue(germ.counts = x, intervals = int, total.seeds = 200,
          method = "dp", k = 10)

#> $`Germination Value`
#> [1] 53.36595
#> 
#> [[2]]
#>    germ.counts intervals Cumulative.germ.counts Cumulative.germ.percent
#> 3           34         3                     34                    17.0
#> 4           40         4                     74                    37.0
#> 5           21         5                     95                    47.5
#> 6           10         6                    105                    52.5
#> 7            4         7                    109                    54.5
#> 8            5         8                    114                    57.0
#> 9            3         9                    117                    58.5
#> 10           5        10                    122                    61.0
#> 11           8        11                    130                    65.0
#> 12           7        12                    137                    68.5
#> 13           7        13                    144                    72.0
#> 14           6        14                    150                    75.0
#> 15           6        15                    156                    78.0
#> 16           4        16                    160                    80.0
#> 17           0        17                    160                    80.0
#> 18           2        18                    162                    81.0
#> 19           0        19                    162                    81.0
#> 20           2        20                    164                    82.0
#>         DGS SumDGSbyN        GV
#> 3  5.666667  5.666667  9.633333
#> 4  9.250000  7.458333 27.595833
#> 5  9.500000  8.138889 38.659722
#> 6  8.750000  8.291667 43.531250
#> 7  7.785714  8.190476 44.638095
#> 8  7.125000  8.012897 45.673512
#> 9  6.500000  7.796769 45.611097
#> 10 6.100000  7.584673 46.266503
#> 11 5.909091  7.398497 48.090230
#> 12 5.708333  7.229481 49.521942
#> 13 5.538462  7.075752 50.945411
#> 14 5.357143  6.932534 51.994006
#> 15 5.200000  6.799262 53.034246
#> 16 5.000000  6.670744 53.365948
#> 17 4.705882  6.539753 52.318022
#> 18 4.500000  6.412268 51.939373
#> 19 4.263158  6.285850 50.915385
#> 20 4.100000  6.164414 50.548194
#> 
#> $testend
#> [1] 16
#> 

# From cumulative germination counts
#----------------------------------------------------------------------------
PeakValue(germ.counts = y, interval = int, total.seeds = 200,
          partial = FALSE)

#> [1] 9.5

GermValue(germ.counts = y, intervals = int, total.seeds = 200,
          partial = FALSE, method = "czabator")

#> $`Germination Value`
#> [1] 38.95
#> 
#> [[2]]
#>    germ.counts intervals Cumulative.germ.counts Cumulative.germ.percent
#> 3           34         3                     34                    17.0
#> 4           40         4                     74                    37.0
#> 5           21         5                     95                    47.5
#> 6           10         6                    105                    52.5
#> 7            4         7                    109                    54.5
#> 8            5         8                    114                    57.0
#> 9            3         9                    117                    58.5
#> 10           5        10                    122                    61.0
#> 11           8        11                    130                    65.0
#> 12           7        12                    137                    68.5
#> 13           7        13                    144                    72.0
#> 14           6        14                    150                    75.0
#> 15           6        15                    156                    78.0
#> 16           4        16                    160                    80.0
#> 17           0        17                    160                    80.0
#> 18           2        18                    162                    81.0
#> 19           0        19                    162                    81.0
#> 20           2        20                    164                    82.0
#>         DGS
#> 3  5.666667
#> 4  9.250000
#> 5  9.500000
#> 6  8.750000
#> 7  7.785714
#> 8  7.125000
#> 9  6.500000
#> 10 6.100000
#> 11 5.909091
#> 12 5.708333
#> 13 5.538462
#> 14 5.357143
#> 15 5.200000
#> 16 5.000000
#> 17 4.705882
#> 18 4.500000
#> 19 4.263158
#> 20 4.100000
#> 

GermValue(germ.counts = y, intervals = int, total.seeds = 200,
          partial = FALSE, method = "dp", k = 10)

#> $`Germination Value`
#> [1] 53.36595
#> 
#> [[2]]
#>    germ.counts intervals Cumulative.germ.counts Cumulative.germ.percent
#> 3           34         3                     34                    17.0
#> 4           40         4                     74                    37.0
#> 5           21         5                     95                    47.5
#> 6           10         6                    105                    52.5
#> 7            4         7                    109                    54.5
#> 8            5         8                    114                    57.0
#> 9            3         9                    117                    58.5
#> 10           5        10                    122                    61.0
#> 11           8        11                    130                    65.0
#> 12           7        12                    137                    68.5
#> 13           7        13                    144                    72.0
#> 14           6        14                    150                    75.0
#> 15           6        15                    156                    78.0
#> 16           4        16                    160                    80.0
#> 17           0        17                    160                    80.0
#> 18           2        18                    162                    81.0
#> 19           0        19                    162                    81.0
#> 20           2        20                    164                    82.0
#>         DGS SumDGSbyN        GV
#> 3  5.666667  5.666667  9.633333
#> 4  9.250000  7.458333 27.595833
#> 5  9.500000  8.138889 38.659722
#> 6  8.750000  8.291667 43.531250
#> 7  7.785714  8.190476 44.638095
#> 8  7.125000  8.012897 45.673512
#> 9  6.500000  7.796769 45.611097
#> 10 6.100000  7.584673 46.266503
#> 11 5.909091  7.398497 48.090230
#> 12 5.708333  7.229481 49.521942
#> 13 5.538462  7.075752 50.945411
#> 14 5.357143  6.932534 51.994006
#> 15 5.200000  6.799262 53.034246
#> 16 5.000000  6.670744 53.365948
#> 17 4.705882  6.539753 52.318022
#> 18 4.500000  6.412268 51.939373
#> 19 4.263158  6.285850 50.915385
#> 20 4.100000  6.164414 50.548194
#> 
#> $testend
#> [1] 16
#>