Germination Time

Compute the following metrics:

MeanGermTime

Mean germination time (T) or Mean length of incubation time (Edmond and Drapala 1958; Czabator 1962; Ellis and Roberts 1980; Labouriau 1983; Ranal and de Santana 2006) .

VarGermTime

Variance of germination time (s2T) (Labouriau 1983; Ranal and de Santana 2006) .

SEGermTime

Standard error of germination time (sT) (Labouriau 1983; Ranal and de Santana 2006) .

CVSEGermTime

Coefficient of variation of the germination time (CVT) (Ranal and de Santana 2006) .

MeanGermTime(germ.counts, intervals, partial = TRUE)

VarGermTime(germ.counts, intervals, partial = TRUE)

SEGermTime(germ.counts, intervals, partial = TRUE)

CVSEGermTime(germ.counts, intervals, partial = TRUE)

Arguments

germ.counts
Germination counts at each time interval. Can be partial or cumulative as specified in the argument partial.

intervals
The time intervals.

partial
logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

Value

For MeanGermTime, the mean germination time value in the same unit of time as specified in the argument intervals.

For VarGermTime, the variance of germination time value as time2.

For SEGermTime, the standard error of germination time in the same unit of time specified in the argument intervals.

For CVSEGermTime, the value of coefficient of variation of the germination time.

Details

MeanGermTime computes the mean germination time according to the following formula (Edmond and Drapala 1958; Czabator 1962; Ellis and Roberts 1980; Labouriau 1983; Ranal and de Santana 2006) .

T = ∑ki=1 NiTi ⁄ ∑ki=1 Ni

Where, Ti is the time from the start of the experiment to the ith observation, Ni is the number of seeds germinated in the ith time (not the accumulated number, but the number correspondent to the ith observation) and k is the last time of germination.

It is the inverse of mean germination rate (V).

T = 1 ⁄ V

It indicates the average length of time required for maximum germination of a seed lot. Lower the T, faster the sample has germinated and reflects seed vigor.

VarGermTime computes the variance of germination time according to the following formula (Labouriau 1983; Ranal and de Santana 2006) .

s2T = ∑ki=1 Ni(Ti − T)2 ⁄ ∑ki=1 Ni − 1

Where, Ti is the time from the start of the experiment to the ith observation, Ni is the number of seeds germinated in the ith time (not the accumulated number, but the number correspondent to the ith observation) and k is the last time of germination.

SEGermTime computes the standard error of germination time (sT) according to the following formula (Labouriau 1983; Ranal and de Santana 2006) .

sT = √[ s2T ⁄ ∑ki=1 Ni ]

Where, Ni is the number of seeds germinated in the ith time (not the accumulated number, but the number correspondent to the ith observation) and k is the last time of germination.

It signifies the accuracy of the calculation of the mean germination time.

CVGermTime computes the coefficient of variation of germination time (CVT) according to the following formula (Ranal and de Santana 2006) .

CVT = √[ s2T ⁄ T ]

This indicates the uniformity of germination and permits comparisons irrespective of the magnitude of mean germination time (T).

References

Czabator FJ (1962). “Germination value: An index combining speed and completeness of pine seed germination.” Forest Science, 8(4), 386--396.

Edmond JB, Drapala WJ (1958). “The effects of temperature, sand and soil, and acetone on germination of okra seed.” Proceedings of the American Society for Horticultural Science, 71, 428--434.

Ellis RH, Roberts EH (1980). “Improved equations for the prediction of seed longevity.” Annals of Botany, 45(1), 13--30.

Labouriau LG (1983). A Germinacao Das Sementes. Organizacao dos Estados Americanos. Programa Regional de Desenvolvimento Cientifico e Tecnologico. Serie de Biologia. Monografia 24.

Ranal MA, de Santana DG (2006). “How and why to measure the germination process?” Brazilian Journal of Botany, 29(1), 1--11.

See also

Examples

x <- c(0, 0, 0, 0, 4, 17, 10, 7, 1, 0, 1, 0, 0, 0)
y <- c(0, 0, 0, 0, 4, 21, 31, 38, 39, 39, 40, 40, 40, 40)
int <- 1:length(x)

# From partial germination counts
#----------------------------------------------------------------------------
MeanGermTime(germ.counts = x, intervals = int)

#> [1] 6.7

VarGermTime(germ.counts = x, intervals = int)

#> [1] 1.446154

SEGermTime(germ.counts = x, intervals = int)

#> [1] 0.1901416

CVSEGermTime(germ.counts = x, intervals = int)

#> [1] 0.1794868

# From cumulative germination counts
#----------------------------------------------------------------------------
MeanGermTime(germ.counts = y, intervals = int, partial = FALSE)

#> [1] 6.7

VarGermTime(germ.counts = y, intervals = int, partial = FALSE)

#> [1] 19.04012

SEGermTime(germ.counts = y, intervals = int, partial = FALSE)

#> [1] 0.2394781

CVSEGermTime(germ.counts = y, intervals = int, partial = FALSE)

#> [1] 0.6512685