Timson's index

Compute George's index, Timson's index or Timson's germination velocity index and it's modifications by Labouriau (Ranal and de Santana, 2006) and Khan and Unger (1984).

TimsonsIndex(germ.counts, intervals, partial = TRUE, total.seeds, max,
  modification = c("none", "labouriau", "khanunger"))

GermRateGeorge(germ.counts, intervals, partial = TRUE, max)

Arguments

germ.counts
Germination counts at each time interval. Can be partial or cumulative as specified in the argument partial.

intervals
The time intervals.

partial
logical. If TRUE, germ.counts is considered as partial and if FALSE, it is considered as cumulative. Default is TRUE.

total.seeds
Total number of seeds.

max
The maximum interval value up to which Timson's index/George's germination rate is to be computed.

modification
The modification to be applied. Either "none", "labouriau" or "khanunger". Default is "none" (see Details).

Value

For TimsonsIndex The value of Timson's index.

For GermRateGeorge The value of George's germination rate.

Details

Timson's index (Timson, 1965) is computed as follows (Brown and Mayer, 1988; Baskin and Baskin, 1998).

∑n = ∑ti=1 Gi

Where, Gi is the cumulative germination percentage in time interval i and t is the total number of time intervals.

For example ten summation (\(\Sigma 10\)) is expressed as follows:

∑ 10 = G1 + G2 + G3 +…+ G10

Where G1, G2, G3, … G10 are the cumulative germination percentage at day 1, 2, 3, …, 10 respectively.

Similarly \(\Sigma 5\) or \(\Sigma 20\) can be estimated. For \(\Sigma 10\), the value can range from 0 (no germination) to 1,000 (100% germination after 24 hours).

It is the progressive total of cumulative germination percentage recorded at specific intervals for a set period of time. It combines onset, rate and total percentage of germination and estimates the area under the cumulative germination percentage curve.

Goodchild and Walker (1971), described the same in terms of partial germination percentage as follows:

∑n = ∑ti=1 gi(t-j)

Where, gi is the germination (not cumulative, but partial germination) in time interval i (i varying from 0 to t) and t is the total number of time intervals and j = i - 1.

Timson's index is similar to the germination rate proposed by George (1961) as follows (Tucker and Wright, 1965; Nichols and Heydecker, 1968).

GR = ∑ti=1 NiKi

Where Ni is the number of seeds germinated by ith interval and Ki is the number of intervals(e.g. days) until the end of the test.

This index uses number of seeds germinated instead of germination percentage.

As Timson's index is useful for comparison only when samples have similar germinabilities or final germination percentage, the following modification was suggested by Labouriau (Ranal and de Santana, 2006).

Tmod = T ⁄ ∑ti=1 gi

Here Timson's index (t) is divided by the sum of partial germination percentages.

Similarly another modification was proposed by Khan and Unger (1984), where Timson's index (t) is divided by number of intervals (t).

Tmod = T ⁄ t

References

George D (1961). “Influence of germination temperature on the expression of post-harvest dormancy in wheat.” Crop Science Abstracts; Western Society of Crop Science Annual Meeting, 1961, 15.

Timson J (1965). “New Method of Recording Germination Data.” Nature, 207(4993), 216.

Tucker H, Wright LN (1965). “Estimating rapidity of germination.” Crop Science, 5(5), 398--399.

Nichols M, Heydecker W (1968). “Two approaches to the study of germination data.” Proceedings of the International Seed Testing Association, 33(3), 531--540.

Goodchild NA, Walker MG (1971). “A method of measuring seed germination in physiological studies.” Annals of Botany, 35(141), 615--621. http://www.jstor.org/stable/42908843.

Khan MA, Ungar IA (1984). “The effect of salinity and temperature on the germination of polymorphic seeds and growth of Atriplex triangularis Willd.” American Journal of Botany, 71(4), 481--489.

Brown RF, Mayer DG (1988). “Representing cumulative germination. 1. A critical analysis of single-value germination indices.” Annals of Botany, 61(2), 117--125.

Baskin CC, Baskin JM (1998). Seeds: Ecology, Biogeography, and Evolution of Dormancy and Germination. Academic Press, San Diego. ISBN 0-12-080260-0.

Ranal MA, de Santana DG (2006). “How and why to measure the germination process?” Brazilian Journal of Botany, 29(1), 1--11.

Examples

x <- c(0, 0, 0, 0, 4, 17, 10, 7, 1, 0, 1, 0, 0, 0)
y <- c(0, 0, 0, 0, 4, 21, 31, 38, 39, 39, 40, 40, 40, 40)
int <- 1:length(x)

# From partial germination counts
#----------------------------------------------------------------------------
# Wihout max specified
TimsonsIndex(germ.counts = x, intervals = int, total.seeds = 50)

#> [1] 664

TimsonsIndex(germ.counts = x, intervals = int, total.seeds = 50,
             modification = "none")

#> [1] 664

TimsonsIndex(germ.counts = x, intervals = int, total.seeds = 50,
             modification = "labouriau")

#> [1] 8.3

TimsonsIndex(germ.counts = x, intervals = int, total.seeds = 50,
             modification = "khanunger")

#> [1] 47.42857

# With max specified
TimsonsIndex(germ.counts = x, intervals = int, total.seeds = 50, max = 10)

#> [1] 344

TimsonsIndex(germ.counts = x, intervals = int, total.seeds = 50,
             max = 10, modification = "none")

#> [1] 344

TimsonsIndex(germ.counts = x, intervals = int, total.seeds = 50,
             max = 10, modification = "labouriau")

#> [1] 4.410256

TimsonsIndex(germ.counts = x, intervals = int, total.seeds = 50,
             max = 10, modification = "khanunger")

#> [1] 24.57143

# Wihout max specified
GermRateGeorge(germ.counts = x, intervals = int)

#> [1] 332

# With max specified
GermRateGeorge(germ.counts = x, intervals = int, max = 10)

#> [1] 172

GermRateGeorge(germ.counts = x, intervals = int, max = 14)

#> [1] 332

# From cumulative germination counts
#----------------------------------------------------------------------------

# Wihout max specified
GermRateGeorge(germ.counts = x, intervals = int, partial = TRUE)

#> [1] 332

# With max specified
GermRateGeorge(germ.counts = x, intervals = int, partial = TRUE, max = 10)

#> [1] 172

GermRateGeorge(germ.counts = x, intervals = int, partial = TRUE, max = 14)

#> [1] 332